[Onboarding] refactor popular artist & artist search results component and re-use the generic ItemLink component
