+++ src/Components/Onboarding/Steps/Artists/ArtistSearchResults.tsx
@@ -1,9 +1,9 @@
 import * as React from "react"
-import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
+import { commitMutation, createFragmentContainer, graphql, QueryRenderer, RelayProp } from "react-relay"
 
 import { RecordSourceSelectorProxy, SelectorData } from "relay-runtime"
 import { ContextConsumer, ContextProps } from "../../../Artsy"
-import SelectableItemContainer from "./SelectableItemContainer"
+import ItemLink from "../../ItemLink"
 
 export interface RelayProps {
   term: string
@@ -12,22 +12,75 @@ export interface RelayProps {
   }
 }
 
-class ArtistSearchResultsContent extends React.Component<RelayProps, null> {
+interface Props extends React.HTMLProps<HTMLAnchorElement>, RelayProps {
+  relay?: RelayProp
+}
+
+class ArtistSearchResultsContent extends React.Component<Props, null> {
   onArtistFollowed(artistId: string, store: RecordSourceSelectorProxy, data: SelectorData): void {
     const suggestedArtist = store.get(data.followArtist.artist.related.suggested.edges[0].node.__id)
 
     const popularArtistsRootField = store.get("client:root:viewer")
     const popularArtists = popularArtistsRootField.getLinkedRecords("match_artist", { term: this.props.term })
-    const updatedPopularArtists = popularArtists.map(popularArtist =>
-      popularArtist.getDataID() === artistId ? suggestedArtist : popularArtist)
+    const updatedPopularArtists = popularArtists.map(
+      popularArtist => (popularArtist.getDataID() === artistId ? suggestedArtist : popularArtist)
+    )
 
     popularArtistsRootField.setLinkedRecords(updatedPopularArtists, "match_artist", { term: this.props.term })
   }
 
+  onFollowedArtist(artist: any) {
+    commitMutation(this.props.relay.environment, {
+      mutation: graphql`
+        mutation ArtistSearchResultsArtistMutation($input: FollowArtistInput!) {
+          followArtist(input: $input) {
+            artist {
+              __id
+              related {
+                suggested(first: 1) {
+                  edges {
+                    node {
+                      id
+                      __id
+                      name
+                      image {
+                        cropped(width: 100, height: 100) {
+                          url
+                        }
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      `,
+      variables: {
+        input: {
+          artist_id: artist.id,
+          unfollow: false,
+        },
+      },
+      updater: (store: RecordSourceSelectorProxy, data: SelectorData) =>
+        this.onArtistFollowed(artist.__id, store, data),
+    })
+  }
+
   render() {
-    return <SelectableItemContainer
-              artists={this.props.viewer.match_artist}
-              onArtistFollowed={this.onArtistFollowed.bind(this)} />
+    const artistItems = this.props.viewer.match_artist.map((artist, index) => (
+      <ItemLink
+        href="#"
+        item={artist}
+        key={index}
+        id={artist.id}
+        name={artist.name}
+        image_url={artist.image && artist.image.cropped.url}
+        onClick={() => this.onFollowedArtist(artist)}
+      />
+    ))
+
+    return <div>{artistItems}</div>
   }
 }
 
@@ -36,7 +89,14 @@ const ArtistSearchResultsContentContainer = createFragmentContainer(
   graphql`
     fragment ArtistSearchResultsContent_viewer on Viewer {
       match_artist(term: $term) {
-        ...SelectableItemContainer_artists
+        id
+        __id
+        name
+        image {
+          cropped(width: 100, height: 100) {
+            url
+          }
+        }
       }
     }
   `
