+++ src/Components/Onboarding/Steps/Artists/PopularArtists.tsx
@@ -1,9 +1,9 @@
 import * as React from "react"
-import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
+import { commitMutation, createFragmentContainer, graphql, QueryRenderer, RelayProp } from "react-relay"
 
 import { RecordSourceSelectorProxy, SelectorData } from "relay-runtime"
 import { ContextConsumer, ContextProps } from "../../../Artsy"
-import SelectableItemContainer from "./SelectableItemContainer"
+import ItemLink from "../../ItemLink"
 
 export interface RelayProps {
   popular_artists: {
@@ -11,22 +11,75 @@ export interface RelayProps {
   }
 }
 
-class PopularArtistsContent extends React.Component<RelayProps, null> {
+interface Props extends React.HTMLProps<HTMLAnchorElement>, RelayProps {
+  relay?: RelayProp
+}
+
+class PopularArtistsContent extends React.Component<Props, null> {
   onArtistFollowed(artistId: string, store: RecordSourceSelectorProxy, data: SelectorData): void {
     const suggestedArtist = store.get(data.followArtist.artist.related.suggested.edges[0].node.__id)
 
     const popularArtistsRootField = store.get("client:root:popular_artists")
     const popularArtists = popularArtistsRootField.getLinkedRecords("artists")
-    const updatedPopularArtists = popularArtists.map(popularArtist =>
-      popularArtist.getDataID() === artistId ? suggestedArtist : popularArtist)
+    const updatedPopularArtists = popularArtists.map(
+      popularArtist => (popularArtist.getDataID() === artistId ? suggestedArtist : popularArtist)
+    )
 
     popularArtistsRootField.setLinkedRecords(updatedPopularArtists, "artists")
   }
 
+  onFollowedArtist(artist: any) {
+    commitMutation(this.props.relay.environment, {
+      mutation: graphql`
+        mutation PopularArtistsFollowArtistMutation($input: FollowArtistInput!) {
+          followArtist(input: $input) {
+            artist {
+              __id
+              related {
+                suggested(first: 1) {
+                  edges {
+                    node {
+                      id
+                      __id
+                      name
+                      image {
+                        cropped(width: 100, height: 100) {
+                          url
+                        }
+                      }
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      `,
+      variables: {
+        input: {
+          artist_id: artist.id,
+          unfollow: false,
+        },
+      },
+      updater: (store: RecordSourceSelectorProxy, data: SelectorData) =>
+        this.onArtistFollowed(artist.__id, store, data),
+    })
+  }
+
   render() {
-    return <SelectableItemContainer
-              artists={this.props.popular_artists.artists}
-              onArtistFollowed={this.onArtistFollowed.bind(this)} />
+    const artistItems = this.props.popular_artists.artists.map((artist, index) => (
+      <ItemLink
+        href="#"
+        item={artist}
+        key={index}
+        id={artist.id}
+        name={artist.name}
+        image_url={artist.image && artist.image.cropped.url}
+        onClick={() => this.onFollowedArtist(artist)}
+      />
+    ))
+
+    return <div>{artistItems}</div>
   }
 }
 
@@ -35,7 +88,14 @@ const PopularArtistContentContainer = createFragmentContainer(
   graphql`
     fragment PopularArtistsContent_popular_artists on PopularArtists {
       artists {
-        ...SelectableItemContainer_artists
+        id
+        __id
+        name
+        image {
+          cropped(width: 100, height: 100) {
+            url
+          }
+        }
       }
     }
   `
